requirements.txt
 pydantic>=1.10.15,
 six,
 pyyaml,
 fastapi==0.109.2,
 uvicorn[standard]==0.27.0.post1,
 transformers==4.25.1,
 python-json-logger==2.0.7,
 torch==2.3.0,
 spacy==3.4.4,
 en-core-web-trf==3.4.1,
 scikit-learn>=1.4.2,
 sentence_transformers==2.2.2,
 hdbscan==0.8.33,
 pandas==2.2.2,
 huggingface-hub>=0.21.4,
 flair==0.13.1,
 seqeval,
 tqdm,
 scipy==1.12.0,
 nltk==3.8.1

